Rename lib to libs
