+++ src/libs/addStorageEventHandler/index.native.js
