+++ src/libs/openURLInNewTab/index.js
