+++ src/libs/openURLInNewTab/index.native.js
